ig_manager/requirements.txt
constructs>=10.0.0,<11.0.0
openai
python-dotenv
pydantic==1.*
langchain
langchain-openai
langchainhub
langsmith
bs4
tiktoken
numexpr
pymongo